move from subscribe -> pipe.map.subscribe for data loader/filter
