+++ src/app/modules/organizations/columns/data-types/data-types-properties.component.ts
@@ -7,7 +7,7 @@ import { MatFormFieldModule } from '@angular/material/form-field'
 import { MatPaginator } from '@angular/material/paginator'
 import { MatSelectModule } from '@angular/material/select'
 import { MatTableDataSource, MatTableModule } from '@angular/material/table'
-import { takeUntil } from 'rxjs'
+import { map, takeUntil } from 'rxjs'
 import { type Column } from '@seed/api/column'
 import { SharedImports } from '@seed/directives'
 import { type UploaderResponse } from '@seed/services/uploader/uploader.types'
@@ -23,7 +23,6 @@ import { DataTypesComponent } from './data-types.component'
 })
 export class DataTypesPropertiesComponent extends DataTypesComponent implements AfterViewInit, OnInit {
   type = 'PropertyState'
-  private _dialog = inject(MatDialog)
   columnTableDataSource = new MatTableDataSource<Column>([])
   columnTableColumns = [
     'display_name',
@@ -32,43 +31,23 @@ export class DataTypesPropertiesComponent extends DataTypesComponent implements 
   @ViewChild(MatPaginator) paginator: MatPaginator
 
   ngOnInit(): void {
-    this._columnService.propertyColumns$.pipe(takeUntil(this._unsubscribeAll$)).subscribe((columns) => {
-      this.columns = columns.sort((a, b) => naturalSort(a.display_name, b.display_name)).filter((c) => c.is_extra_data)
-      for (const c of this.columns) {
-        this.dataTypesForm.addControl(`${c.id}`, new FormControl((c.data_type), [Validators.required]))
-      }
-      if (this.columns.length > 0) {
-        this.isLoading = false
-      }
-      this.columnTableDataSource.data = this.columns
-    })
+    this._columnService.propertyColumns$.pipe(takeUntil(this._unsubscribeAll$)).pipe(
+      map((columns) => {
+        this.columns = columns.sort((a, b) => naturalSort(a.display_name, b.display_name)).filter((c) => c.is_extra_data)
+        for (const c of this.columns) {
+          this.dataTypesForm.addControl(`${c.id}`, new FormControl((c.data_type), [Validators.required]))
+        }
+        if (this.columns.length > 0) {
+          this.isLoading = false
+        }
+        this.columnTableDataSource.data = this.columns
+      })
+    ).subscribe()
   }
 
   ngAfterViewInit(): void {
     this.columnTableDataSource.paginator = this.paginator
   }
 
-  save(): void {
-    const changes = {}
-    for (const column of this.columns) {
-      const setting = this.dataTypesForm.get(`${column.id}`).value
-      if (setting !== column.data_type) {
-        changes[column.id] = { data_type: setting }
-      }
-    }
-    if (Object.keys(changes).length > 0) {
-      this._columnService.updateMultipleColumns(this.columns[0].organization_id, this.type, changes).subscribe((response: UploaderResponse) => {
-        const dialogRef = this._dialog.open(UpdateModalComponent, {
-          width: '40rem',
-          data: { progressResponse: response },
-        })
-        dialogRef
-          .afterClosed()
-          .pipe(
-            takeUntil(this._unsubscribeAll$),
-          )
-          .subscribe()
-      })
-    }
-  }
+
 }
