+++ src/app/modules/organizations/columns/data-types/data-types.component.ts
@@ -1,8 +1,11 @@
 import { Component, inject, type OnDestroy, ViewEncapsulation } from '@angular/core'
 import { FormGroup } from '@angular/forms'
-import { Subject } from 'rxjs'
+import { MatDialog } from '@angular/material/dialog'
+import { Subject, takeUntil } from 'rxjs'
 import { type Column, ColumnService } from '@seed/api/column'
 import { SharedImports } from '@seed/directives'
+import { UpdateModalComponent } from '../modal/update-modal.component'
+import { type UploaderResponse } from '@seed/services/uploader'
 import { DataTypes } from './data-types.constants'
 
 @Component({
@@ -15,6 +18,7 @@ import { DataTypes } from './data-types.constants'
 export class DataTypesComponent implements OnDestroy {
   protected _columnService = inject(ColumnService)
   protected readonly _unsubscribeAll$ = new Subject<void>()
+  private _dialog = inject(MatDialog)
   columns: Column[]
   type: string
   dataTypesForm = new FormGroup({})
@@ -24,5 +28,29 @@ export class DataTypesComponent implements OnDestroy {
   ngOnDestroy(): void {
     this._unsubscribeAll$.next()
     this._unsubscribeAll$.complete()
+  }
+
+  save(): void {
+    const changes = {}
+    for (const column of this.columns) {
+      const setting = this.dataTypesForm.get(`${column.id}`).value
+      if (setting !== column.data_type) {
+        changes[column.id] = { data_type: setting }
+      }
+    }
+    if (Object.keys(changes).length > 0) {
+      this._columnService.updateMultipleColumns(this.columns[0].organization_id, this.type, changes).subscribe((response: UploaderResponse) => {
+        const dialogRef = this._dialog.open(UpdateModalComponent, {
+          width: '40rem',
+          data: { progressResponse: response },
+        })
+        dialogRef
+          .afterClosed()
+          .pipe(
+            takeUntil(this._unsubscribeAll$),
+          )
+          .subscribe()
+      })
+    }
   }
 }
